requirements.txt
click==6.7
Cython==0.25.2
fastTextpy==0.0.3
Flask==0.11.1
Flask-SocketIO==2.9.2
gevent==1.2.2
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.12.1
pybind11==2.2.1
python-engineio==2.0.1
python-socketio==1.8.3
six==1.11.0
Werkzeug==0.12.2
